LayoutTests/inspector/version-controller-expected.txt

Tests inspector version controller.


Running: testMethodsToRunToUpdateVersion
Testing methods to run to upgrade from 0 to 0.
Methods to run: []

Testing methods to run to upgrade from 0 to 1.
Methods to run: ["_updateVersionFrom0To1"]

Testing methods to run to upgrade from 0 to 2.
Methods to run: ["_updateVersionFrom0To1","_updateVersionFrom1To2"]

Testing methods to run to upgrade from 0 to 3.
Methods to run: ["_updateVersionFrom0To1","_updateVersionFrom1To2","_updateVersionFrom2To3"]

Testing methods to run to upgrade from 1 to 1.
Methods to run: []

Testing methods to run to upgrade from 1 to 2.
Methods to run: ["_updateVersionFrom1To2"]

Testing methods to run to upgrade from 1 to 3.
Methods to run: ["_updateVersionFrom1To2","_updateVersionFrom2To3"]

Testing methods to run to upgrade from 2 to 2.
Methods to run: []

Testing methods to run to upgrade from 2 to 3.
Methods to run: ["_updateVersionFrom2To3"]


Running: testClearBreakpointsWhenTooMany
Starting test with 0 breakpoints.
Number of breakpoints left in the setting after the test: 0.

Starting test with 1 breakpoints.
Number of breakpoints left in the setting after the test: 1.

Starting test with 2 breakpoints.
Number of breakpoints left in the setting after the test: 2.

Starting test with 499999 breakpoints.
Number of breakpoints left in the setting after the test: 499999.

Starting test with 500000 breakpoints.
Number of breakpoints left in the setting after the test: 500000.

Starting test with 500001 breakpoints.
Number of breakpoints left in the setting after the test: 0.

Starting test with 1000000 breakpoints.
Number of breakpoints left in the setting after the test: 0.
